Initial draft version
This commit serves as an initial draft version.  It did not run yet
and there is a high chance that the SPARQL queries fail.  However, it
is a starting point upon which we can build.

Do note that the URIs which are currently used are highly subject to
change.  Feedback is much appreciated.