Introduction of muAccount:AggregatedSSNAccess
In the data model, we introduce a new InformationResource an account
can be entitled to.
This will serve the purpose of splitting out in data a normal vendor
account, which can only find personUri for one bestuurseeneheide, and
a 'super' vendor account (like Mijn Burgerprofiel) which can fetch a
personUri over ALL bestuurseenheden.